+++ aws-android-sdk-sdb/src/main/java/com/amazonaws/services/simpledb/util/SimpleDBUtils.java
@@ -12,6 +12,7 @@
  * express or implied. See the License for the specific language governing
  * permissions and limitations under the License.
  */
+
 package com.amazonaws.services.simpledb.util;
 
 import java.text.ParseException;
@@ -31,15 +32,13 @@ public class SimpleDBUtils {
      */
     private static String dateFormat = "yyyy-MM-dd'T'HH:mm:ss.SSSZ";
 
-
     /**
      * Encodes positive integer value into a string by zero-padding number up to
      * the specified number of digits.
      *
-     * @param number
-     *            positive integer to be encoded
-     * @param maxNumDigits
-     *            maximum number of digits in the largest value in the data set
+     * @param number positive integer to be encoded
+     * @param maxNumDigits maximum number of digits in the largest value in the
+     *            data set
      * @return string representation of the zero-padded integer
      */
     public static String encodeZeroPadding(int number, int maxNumDigits) {
@@ -57,10 +56,9 @@ public class SimpleDBUtils {
      * Encodes positive long value into a string by zero-padding the value up to
      * the specified number of digits.
      *
-     * @param number
-     *            positive long to be encoded
-     * @param maxNumDigits
-     *            maximum number of digits in the largest value in the data set
+     * @param number positive long to be encoded
+     * @param maxNumDigits maximum number of digits in the largest value in the
+     *            data set
      * @return string representation of the zero-padded long
      */
     public static String encodeZeroPadding(long number, int maxNumDigits) {
@@ -78,11 +76,9 @@ public class SimpleDBUtils {
      * Encodes positive float value into a string by zero-padding number up to
      * the specified number of digits
      *
-     * @param number
-     *            positive float value to be encoded
-     * @param maxNumDigits
-     *            maximum number of digits preceding the decimal point in the
-     *            largest value in the data set
+     * @param number positive float value to be encoded
+     * @param maxNumDigits maximum number of digits preceding the decimal point
+     *            in the largest value in the data set
      * @return string representation of the zero-padded float value
      */
     public static String encodeZeroPadding(float number, int maxNumDigits) {
@@ -101,8 +97,7 @@ public class SimpleDBUtils {
     /**
      * Decodes zero-padded positive integer value from the string representation
      *
-     * @param value
-     *            zero-padded string representation of the integer
+     * @param value zero-padded string representation of the integer
      * @return original integer value
      */
     public static int decodeZeroPaddingInt(String value) {
@@ -112,8 +107,7 @@ public class SimpleDBUtils {
     /**
      * Decodes a zero-padded positive long value from the string representation
      *
-     * @param value
-     *            zero-padded string representation of the long
+     * @param value zero-padded string representation of the long
      * @return original long value
      */
     public static long decodeZeroPaddingLong(String value) {
@@ -123,8 +117,7 @@ public class SimpleDBUtils {
     /**
      * Decodes zero-padded positive float value from the string representation
      *
-     * @param value
-     *            zero-padded string representation of the float value
+     * @param value zero-padded string representation of the float value
      * @return original float value
      */
     public static float decodeZeroPaddingFloat(String value) {
@@ -136,14 +129,11 @@ public class SimpleDBUtils {
      * number up to the specified number of digits. Use this encoding method if
      * the data range set includes both positive and negative values.
      *
-     * @param number
-     *            integer to be encoded
-     * @param maxNumDigits
-     *            maximum number of digits in the largest absolute value in the
-     *            data set
-     * @param offsetValue
-     *            offset value, has to be greater than absolute value of any
-     *            negative number in the data set.
+     * @param number integer to be encoded
+     * @param maxNumDigits maximum number of digits in the largest absolute
+     *            value in the data set
+     * @param offsetValue offset value, has to be greater than absolute value of
+     *            any negative number in the data set.
      * @return string representation of the integer
      */
     public static String encodeRealNumberRange(int number, int maxNumDigits,
@@ -165,14 +155,11 @@ public class SimpleDBUtils {
      * encoding method if the data set includes both positive and negative
      * values.
      *
-     * @param number
-     *            positive or negative long value to be encoded
-     * @param maxNumDigits
-     *            maximum number of digits in the largest absolute value in the
-     *            data set
-     * @param offsetValue
-     *            offset value, has to be greater than absolute value of any
-     *            negative number in the data set.
+     * @param number positive or negative long value to be encoded
+     * @param maxNumDigits maximum number of digits in the largest absolute
+     *            value in the data set
+     * @param offsetValue offset value, has to be greater than absolute value of
+     *            any negative number in the data set.
      * @return string representation of the long
      */
     public static String encodeRealNumberRange(long number, int maxNumDigits,
@@ -193,29 +180,27 @@ public class SimpleDBUtils {
      * number up to the specified number of digits. Use this encoding method if
      * the data range set includes both positive and negative values.
      *
-     * @param number
-     *            float to be encoded
-     * @param maxDigitsLeft
-     *            maximum number of digits left of the decimal point in the
-     *            largest absolute value in the data set
-     * @param maxDigitsRight
-     *            maximum number of digits right of the decimal point in the
-     *            largest absolute value in the data set, i.e. precision
-     * @param offsetValue
-     *            offset value, has to be greater than absolute value of any
-     *            negative number in the data set.
+     * @param number float to be encoded
+     * @param maxDigitsLeft maximum number of digits left of the decimal point
+     *            in the largest absolute value in the data set
+     * @param maxDigitsRight maximum number of digits right of the decimal point
+     *            in the largest absolute value in the data set, i.e. precision
+     * @param offsetValue offset value, has to be greater than absolute value of
+     *            any negative number in the data set.
      * @return string representation of the integer
      */
     public static String encodeRealNumberRange(float number, int maxDigitsLeft,
             int maxDigitsRight, int offsetValue) {
         int shiftMultiplier = (int) Math.pow(10, maxDigitsRight);
-        long shiftedNumber = (long) Math.round((double)number * shiftMultiplier);
+        long shiftedNumber = Math.round((double) number * shiftMultiplier);
         long shiftedOffset = offsetValue * shiftMultiplier;
         long offsetNumber = shiftedNumber + shiftedOffset;
 
         if (offsetNumber < 0)
         {
-            throw new IllegalArgumentException("OffsetNumber[" + offsetNumber + "] is negative - Number[" + number + "], maxDigitsLeft[" + maxDigitsLeft + "], maxDigitsRight[" + maxDigitsRight + "], offsetValue[" + offsetValue + "]");
+            throw new IllegalArgumentException("OffsetNumber[" + offsetNumber
+                    + "] is negative - Number[" + number + "], maxDigitsLeft[" + maxDigitsLeft
+                    + "], maxDigitsRight[" + maxDigitsRight + "], offsetValue[" + offsetValue + "]");
         }
 
         String longString = Long.toString(offsetNumber);
@@ -224,7 +209,9 @@ public class SimpleDBUtils {
 
         if (numZeroes < 0)
         {
-            throw new IllegalArgumentException("Number[" + number + "] has too many digits - maxDigitsLeft[" + maxDigitsLeft + "], maxDigitsRight[" + maxDigitsRight + "], offsetValue[" + offsetValue + "]");
+            throw new IllegalArgumentException("Number[" + number
+                    + "] has too many digits - maxDigitsLeft[" + maxDigitsLeft
+                    + "], maxDigitsRight[" + maxDigitsRight + "], offsetValue[" + offsetValue + "]");
         }
 
         StringBuffer strBuffer = new StringBuffer(numZeroes + longString.length());
@@ -239,10 +226,8 @@ public class SimpleDBUtils {
      * Decodes integer value from the string representation that was created by
      * using encodeRealNumberRange(..) function.
      *
-     * @param value
-     *            string representation of the integer value
-     * @param offsetValue
-     *            offset value that was used in the original encoding
+     * @param value string representation of the integer value
+     * @param offsetValue offset value that was used in the original encoding
      * @return original integer value
      */
     public static int decodeRealNumberRangeInt(String value, int offsetValue) {
@@ -254,45 +239,39 @@ public class SimpleDBUtils {
      * Decodes a long value from the string representation that was created by
      * using encodeRealNumberRange(..) function.
      *
-     * @param value
-     *            string representation of the long value
-     * @param offsetValue
-     *            offset value that was used in the original encoding
+     * @param value string representation of the long value
+     * @param offsetValue offset value that was used in the original encoding
      * @return original long value
      */
     public static long decodeRealNumberRangeLong(String value, long offsetValue) {
         long offsetNumber = Long.parseLong(value, 10);
-        return (long) (offsetNumber - offsetValue);
+        return offsetNumber - offsetValue;
     }
 
     /**
      * Decodes float value from the string representation that was created by
      * using encodeRealNumberRange(..) function.
      *
-     * @param value
-     *            string representation of the integer value
-     * @param maxDigitsRight
-     *            maximum number of digits left of the decimal point in the
-     *            largest absolute value in the data set (must be the same as
-     *            the one used for encoding).
-     * @param offsetValue
-     *            offset value that was used in the original encoding
+     * @param value string representation of the integer value
+     * @param maxDigitsRight maximum number of digits left of the decimal point
+     *            in the largest absolute value in the data set (must be the
+     *            same as the one used for encoding).
+     * @param offsetValue offset value that was used in the original encoding
      * @return original float value
      */
     public static float decodeRealNumberRangeFloat(String value,
             int maxDigitsRight, int offsetValue) {
         long offsetNumber = Long.parseLong(value, 10);
         int shiftMultiplier = (int) Math.pow(10, maxDigitsRight);
-        double tempVal = (double) (offsetNumber - offsetValue * shiftMultiplier);
-        return (float) (tempVal / (double) (shiftMultiplier));
+        double tempVal = offsetNumber - offsetValue * shiftMultiplier;
+        return (float) (tempVal / (shiftMultiplier));
     }
 
     /**
      * Encodes date value into string format that can be compared
      * lexicographically
      *
-     * @param date
-     *            date value to be encoded
+     * @param date date value to be encoded
      * @return string representation of the date value
      */
     public static String encodeDate(Date date) {
@@ -307,8 +286,7 @@ public class SimpleDBUtils {
      * Decodes date value from the string representation created using
      * encodeDate(..) function.
      *
-     * @param value
-     *            string representation of the date value
+     * @param value string representation of the date value
      * @return original date value
      */
     public static Date decodeDate(String value) throws ParseException {
@@ -322,10 +300,8 @@ public class SimpleDBUtils {
      * Quotes and escapes a list of values so that they can be used in a
      * SimpleDB query.
      *
-     * @param values
-     *            The collection of attribute values that will be quoted,
+     * @param values The collection of attribute values that will be quoted,
      *            escaped, and included in the returned string list.
-     *
      * @return A string representation of the list of specified values, with
      *         individual values properly quoted and escaped.
      */
@@ -334,7 +310,8 @@ public class SimpleDBUtils {
 
         boolean first = true;
         for (String s : values) {
-            if (!first) sb.append(",");
+            if (!first)
+                sb.append(",");
             first = false;
             sb.append(quoteValue(s));
         }
@@ -346,41 +323,37 @@ public class SimpleDBUtils {
      * Quotes and escapes an attribute value by wrapping it with single quotes
      * and escaping any single quotes inside the value.
      *
-     * @param value
-     *            The attribute value to quote and escape.
-     *
+     * @param value The attribute value to quote and escape.
      * @return The properly quoted and escaped attribute value, ready to be used
      *         in a SimpleDB select query.
      */
     public static String quoteValue(String value) {
-        return "'" + replaceChar( value, "'", "''" ) + "'";
+        return "'" + replaceChar(value, "'", "''") + "'";
     }
 
     /**
      * Quotes and escapes an attribute name or domain name by wrapping it with
      * backticks and escaping any backticks inside the name.
      *
-     * @param name
-     *            The attribute name or domain name to quote and escape.
-     *
+     * @param name The attribute name or domain name to quote and escape.
      * @return The properly quoted and escaped attribute name or domain name,
      *         ready to be used in a SimpleDB select query.
      */
     public static String quoteName(String name) {
-        return "`" + replaceChar( name, "`", "``" ) + "`";
+        return "`" + replaceChar(name, "`", "``") + "`";
     }
 
-    protected static String replaceChar( String value, String termToFind, String replacementTerm ) {
-        StringBuilder buffer = new StringBuilder( value );
+    protected static String replaceChar(String value, String termToFind, String replacementTerm) {
+        StringBuilder buffer = new StringBuilder(value);
 
         int searchIndex = 0;
-        while ( searchIndex < buffer.length() ) {
-            searchIndex = buffer.indexOf( termToFind, searchIndex );
-            if ( searchIndex == -1 ) {
+        while (searchIndex < buffer.length()) {
+            searchIndex = buffer.indexOf(termToFind, searchIndex);
+            if (searchIndex == -1) {
                 break;
             }
             else {
-                buffer.replace( searchIndex, searchIndex + termToFind.length(), replacementTerm );
+                buffer.replace(searchIndex, searchIndex + termToFind.length(), replacementTerm);
                 searchIndex += replacementTerm.length();
             }
         }
